[Build and Feat] Install TanStack Query and add search function
